Nia Brown Final Project | CSC 496-001
To see the full potential of my code, please check my readME file to view the screen recordings and images that highlight how my project works. Unfortunately, when trying to run locally, the live server freezes, not showing the full potential of my code and making it look "broken". Again, the videos in my readME file highlight the full potential of my project.